requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.7.0
asyncpg==0.30.0
attrs==24.3.0
bcrypt==4.0.1
billiard==4.2.1
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
celery==5.5.3
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crashtest==0.4.1
cryptography==44.0.0
dataclasses-json==0.6.7
deepgram-captions==1.2.0
deepgram-sdk==3.7.7
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
dulwich==0.22.7
ecdsa==0.19.0
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsrs==6.1.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
gzip-stream==1.2.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jwt==1.3.1
keyring==25.6.0
kombu==5.5.4
langchain==0.3.13
langchain-anthropic==0.3.8
langchain-core==0.3.74
langchain-deepseek==0.1.4
langchain-google-genai==2.1.9
langchain-openai==0.3.30
langchain-text-splitters==0.3.4
langsmith==0.4.13
lara-sdk==1.4.5
loguru==0.7.3
magic-filter==1.0.12
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.1
openai==1.99.9
openpyxl==3.1.5
orjson==3.10.12
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.1
passlib==1.7.4
pillow==12.0.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.6.0
poetry==2.0.1
poetry-core==2.0.1
prompt_toolkit==3.0.51
propcache==0.2.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.23.4
Pygments==2.19.1
pymongo==4.10.1
PyPDF2==3.0.1
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
python-telegram-bot==22.5
pytubefix==8.8.2
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
RapidFuzz==3.11.0
redis==6.4.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
selenium==4.37.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.41.3
tenacity==9.0.0
tiktoken==0.9.0
tomlkit==0.13.2
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
trove-classifiers==2025.1.10.15
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
virtualenv==20.28.1
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.9.0
websockets==13.1
wsproto==1.2.0
xattr==1.1.4
yarl==1.18.3
zstandard==0.23.0
